chore(db): migrate from room to sqldelight

+++ component-chatapi-db/src/androidMain/kotlin/fr/outadoc/justchatting/component/chatapi/db/RecentEmotesConfig.kt
@@ -1,3 +1,3 @@
 package fr.outadoc.justchatting.component.chatapi.db
 
-const val MaxRecentEmotes: Int = 50
+const val MaxRecentEmotes: Long = 50
